requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.44
botocore==1.19.44
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
filelock==3.0.12
gevent==20.12.1
greenlet==1.0a1
gunicorn==20.0.4
heroku==0.1.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.6.1
parsel==1.6.0
pbr==5.5.0
Pillow==8.0.1
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pygame==1.9.6
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.25.1
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.2
txaio==20.4.1
urllib3==1.26.2
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
w3lib==1.22.0
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.2.0
